createAsyncIterable util
